fix: se arreglaron mas cosas

+++ Models/Dtos/Monthly Follow-Ups-Dtos/Habits-Dtos/RetrieveResponsesHabitsDto.cs
@@ -12,53 +12,53 @@ namespace AppVidaSana.Models.Dtos.Monthly_Follow_Ups_Dtos.Habits_Dtos
 
         [JsonRequired] public TimeOnly answerQuestion1 { get; set; }
 
-        [JsonRequired] public int answerQuestion2 { get; set; }
+        [JsonRequired] public byte answerQuestion2 { get; set; }
 
         [JsonRequired] public TimeOnly answerQuestion3 { get; set; }
 
         [JsonRequired] public int answerQuestion4 { get; set; }
 
-        [JsonRequired] public int answerQuestion5a { get; set; } 
+        [JsonRequired] public byte answerQuestion5a { get; set; }
 
-        [JsonRequired] public int answerQuestion5b { get; set; } 
+        [JsonRequired] public byte answerQuestion5b { get; set; }
 
-        [JsonRequired] public int answerQuestion5c { get; set; } 
+        [JsonRequired] public byte answerQuestion5c { get; set; }
 
-        [JsonRequired] public int answerQuestion5d { get; set; } 
+        [JsonRequired] public byte answerQuestion5d { get; set; }
 
-        [JsonRequired] public int answerQuestion5e { get; set; } 
+        [JsonRequired] public byte answerQuestion5e { get; set; }
 
-        [JsonRequired] public int answerQuestion5f { get; set; } 
+        [JsonRequired] public byte answerQuestion5f { get; set; }
 
-        [JsonRequired] public int answerQuestion5g { get; set; } 
+        [JsonRequired] public byte answerQuestion5g { get; set; }
 
-        [JsonRequired] public int answerQuestion5h { get; set; } 
+        [JsonRequired] public byte answerQuestion5h { get; set; }
 
-        [JsonRequired] public int answerQuestion5i { get; set; } 
+        [JsonRequired] public byte answerQuestion5i { get; set; }
 
-        [JsonRequired] public int answerQuestion5j { get; set; } 
+        [JsonRequired] public byte answerQuestion5j { get; set; }
 
-        [JsonRequired] public int answerQuestion6 { get; set; } 
+        [JsonRequired] public byte answerQuestion6 { get; set; }
 
-        [JsonRequired] public int answerQuestion7 { get; set; } 
+        [JsonRequired] public byte answerQuestion7 { get; set; }
 
-        [JsonRequired] public int answerQuestion8 { get; set; } 
+        [JsonRequired] public byte answerQuestion8 { get; set; }
 
-        [JsonRequired] public int answerQuestion9 { get; set; } 
+        [JsonRequired] public byte answerQuestion9 { get; set; }
 
-        [JsonRequired] public int resultComponent1 { get; set; }
+        [JsonRequired] public byte resultComponent1 { get; set; }
 
-        [JsonRequired] public int resultComponent2 { get; set; }
+        [JsonRequired] public byte resultComponent2 { get; set; }
 
-        [JsonRequired] public int resultComponent3 { get; set; }
+        [JsonRequired] public byte resultComponent3 { get; set; }
 
-        [JsonRequired] public int resultComponent4 { get; set; }
+        [JsonRequired] public byte resultComponent4 { get; set; }
 
-        [JsonRequired] public int resultComponent5 { get; set; }
+        [JsonRequired] public byte resultComponent5 { get; set; }
 
-        [JsonRequired] public int resultComponent6 { get; set; }
+        [JsonRequired] public byte resultComponent6 { get; set; }
 
-        [JsonRequired] public int resultComponent7 { get; set; }
+        [JsonRequired] public byte resultComponent7 { get; set; }
 
         [JsonRequired] public int globalClassification { get; set; }
