+++ Models/Dtos/Monthly Follow-Ups-Dtos/Habits-Dtos/SaveResultsDto.cs
@@ -10,19 +10,19 @@ namespace AppVidaSana.Models.Dtos.Monthly_Follow_Ups_Dtos.Habits_Dtos
 
         [JsonRequired] public int year { get; set; }
 
-        [JsonRequired] public int resultComponent1 { get; set; }
+        [JsonRequired] public byte resultComponent1 { get; set; }
 
-        [JsonRequired] public int resultComponent2 { get; set; }
+        [JsonRequired] public byte resultComponent2 { get; set; }
 
-        [JsonRequired] public int resultComponent3 { get; set; }
+        [JsonRequired] public byte resultComponent3 { get; set; }
 
-        [JsonRequired] public int resultComponent4 { get; set; }
+        [JsonRequired] public byte resultComponent4 { get; set; }
 
-        [JsonRequired] public int resultComponent5 { get; set; }
+        [JsonRequired] public byte resultComponent5 { get; set; }
 
-        [JsonRequired] public int resultComponent6 { get; set; }
+        [JsonRequired] public byte resultComponent6 { get; set; }
 
-        [JsonRequired] public int resultComponent7 { get; set; }
+        [JsonRequired] public byte resultComponent7 { get; set; }
 
         [JsonRequired] public int globalClassification { get; set; }
