+++ Models/Monthly Follow-Ups/MFUsHabits.cs
@@ -22,7 +22,7 @@ namespace AppVidaSana.Models.Seguimientos_Mensuales
         public TimeOnly answerQuestion1 { get; set; }
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 2 es obligatorio")]
-        public int answerQuestion2 { get; set; }
+        public byte answerQuestion2 { get; set; }
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 3 es obligatoria")]
         public TimeOnly answerQuestion3 { get; set; }
@@ -31,46 +31,46 @@ namespace AppVidaSana.Models.Seguimientos_Mensuales
         public int answerQuestion4 { get; set; }
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 5a es obligatorio")]
-        public int answerQuestion5a { get; set; }
+        public byte answerQuestion5a { get; set; }
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 5b es obligatorio")]
-        public int answerQuestion5b { get; set; }
+        public byte answerQuestion5b { get; set; }
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 5c es obligatorio")]
-        public int answerQuestion5c { get; set; } 
+        public byte answerQuestion5c { get; set; } 
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 5d es obligatorio")]
-        public int answerQuestion5d { get; set; } 
+        public byte answerQuestion5d { get; set; } 
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 5e es obligatorio")]
-        public int answerQuestion5e { get; set; } 
+        public byte answerQuestion5e { get; set; } 
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 5f es obligatorio")]
-        public int answerQuestion5f { get; set; }
+        public byte answerQuestion5f { get; set; }
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 5g es obligatorio")]
-        public int answerQuestion5g { get; set; } 
+        public byte answerQuestion5g { get; set; } 
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 5h es obligatorio")]
-        public int answerQuestion5h { get; set; } 
+        public byte answerQuestion5h { get; set; } 
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 5i es obligatorio")]
-        public int answerQuestion5i { get; set; }
+        public byte answerQuestion5i { get; set; }
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 5j es obligatorio")]
-        public int answerQuestion5j { get; set; }
+        public byte answerQuestion5j { get; set; }
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 6 es obligatorio")]
-        public int answerQuestion6 { get; set; }
+        public byte answerQuestion6 { get; set; }
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 7 es obligatorio")]
-        public int answerQuestion7 { get; set; } 
+        public byte answerQuestion7 { get; set; } 
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 8 es obligatorio")]
-        public int answerQuestion8 { get; set; } 
+        public byte answerQuestion8 { get; set; } 
 
         [Required(ErrorMessage = "El campo respuesta de la pregunta 9 es obligatorio")]
-        public int answerQuestion9 { get; set; } 
+        public byte answerQuestion9 { get; set; } 
 
         public Account? account { get; set; }
